LayoutTests/imported/w3c/web-platform-tests/html/semantics/embedded-content/the-iframe-element/iframe-network-error.sub-expected.txt

Blocked access to external URL http://nonexistent.localhost/
CONSOLE MESSAGE: Refused to display 'http://localhost:8800/html/semantics/embedded-content/resources/not-embeddable.html' in a frame because it set 'X-Frame-Options' to 'deny'.
Blocked access to external URL http://nonexistent.localhost/
CONSOLE MESSAGE: Refused to display 'http://localhost:8800/html/semantics/embedded-content/resources/not-embeddable.html' in a frame because it set 'X-Frame-Options' to 'deny'.


PASS new iframe: nonexistent host
PASS new iframe: X-Frame-Options prevents embedding
PASS navigating an existing iframe: nonexistent host
PASS navigating an existing iframe: X-Frame-Options prevents embedding
